+++ publish.sh
@@ -264,29 +264,16 @@ else
 echo "SKIPPING $dir (unchanged)"
 fi
 
-echo "Initialize and update git submodules"
-git submodule init
-git submodule update
-
-dir=submodule-aws-qnabot-plugins
-if haschanged $dir; then
-echo "PACKAGING $dir"
-pushd $dir
-./publish.sh $BUCKET $PREFIX_AND_VERSION/aws-qnabot-plugins || exit 1
-popd
-update_checksum $dir
-else
-echo "SKIPPING $dir (unchanged)"
-fi
-
 dir=submodule-aws-qnabot
 if haschanged $dir; then
 echo "PACKAGING $dir"
 git submodule init
 git submodule update
-echo "Applying patch files to simplify UX by removing some QnABot options not needed for LCA"
-# LCA customizations
-cp -v ./patches/qnabot/Makefile $dir/Makefile
+# lma customizations
+echo "Applying patch files to remove unused KMS keys from QnABot and customize designer settings page"
+cp -v ./patches/qnabot/templates_examples_examples_index.js $dir/source/templates/examples/examples/index.js
+cp -v ./patches/qnabot/templates_examples_extensions_index.js $dir/source/templates/examples/extensions/index.js
+cp -v ./patches/qnabot/website_js_lib_store_api_actions_settings.js $dir/source/website/js/lib/store/api/actions/settings.js
 echo "modify QnABot version string from 'N.N.N' to 'N.N.N-LCA'"
 # Detection of differences. sed varies betwen GNU sed and BSD sed
 if sed --version 2>/dev/null | grep -q GNU; then # GNU sed
@@ -308,6 +295,10 @@ cat > config.json <<_EOF
 _EOF
 npm install
 npm run build || exit 1
+# Rename OpenbsearchDomain resource in template to force resource replacement during upgrade/downgrade
+# If the resource name is not changed, then CloudFomration does an inline upgrade from OpenSearch 1.3 to 2.1, but this upgrade cannot be reversed
+# which can create a problem with ROLLBACK if there is a stack failure during the upgrade.
+cat ./build/templates/master.json | sed -e "s%OpensearchDomain%LMAQnaBotOpensearchDomain%g" > ./build/templates/qnabot-main.json
 aws s3 sync ./build/ s3://${BUCKET}/${PREFIX_AND_VERSION}/aws-qnabot/ --delete 
 popd
 update_checksum $dir